new/requirements.txt
# Self-Evolving HFT Trading System Dependencies

# Core numerical
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0

# Machine Learning / RL
torch>=2.0.0
scikit-learn>=1.3.0

# Statistics
hmmlearn>=0.3.0
arch>=6.0.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Binance API
python-binance>=1.0.1
aiohttp>=3.8.0
websockets>=12.0

# Visualization (optional)
matplotlib>=3.7.0
seaborn>=0.12.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Utilities
tqdm>=4.65.0
pathlib>=1.0.1
